+++ packages/axios-extension/src/abap/adt-catalog/generators/types.ts
@@ -1,0 +1,8 @@
+export interface GeneratorEntry {
+    id: string;
+    link: {
+        href: string;
+        rel: string;
+        type: string;
+    }[];
+}