+++ packages/preview-middleware-client/test/unit/adp/init-dialogs.test.ts
@@ -3,16 +3,19 @@ import { RTAOptions } from 'sap/ui/rta/RuntimeAuthoring';
 import type ElementOverlay from 'sap/ui/dt/ElementOverlay';
 import type RuntimeAuthoring from 'sap/ui/rta/RuntimeAuthoring';
 
+import FlUtils from 'sap/ui/fl/Utils';
 import { sapMock } from 'mock/window';
 import Utils from 'mock/sap/ui/fl/Utils';
 import Fragment from 'mock/sap/ui/core/Fragment';
 import Controller from 'mock/sap/ui/core/mvc/Controller';
 import RuntimeAuthoringMock from 'mock/sap/ui/rta/RuntimeAuthoring';
 
-import { DialogNames, handler, initDialogs, isFragmentCommandEnabled } from '../../../src/adp/init-dialogs';
+import { DialogNames, handler, initDialogs, isFragmentCommandEnabled, isControllerExtensionEnabled } from '../../../src/adp/init-dialogs';
 import AddFragment from '../../../src/adp/controllers/AddFragment.controller';
 import ControllerExtension from '../../../src/adp/controllers/ControllerExtension.controller';
 import ExtensionPoint from '../../../src/adp/controllers/ExtensionPoint.controller';
+import ElementOverlay from 'sap/ui/dt/ElementOverlay';
+
 
 describe('Dialogs', () => {
     describe('initDialogs', () => {
@@ -28,7 +31,7 @@ describe('Dialogs', () => {
                     addMenuItem: addMenuItemSpy
                 }
             });
-            initDialogs(rtaMock as unknown as RuntimeAuthoring);
+            initDialogs(rtaMock as unknown as RuntimeAuthoring, []);
             expect(addMenuItemSpy).toHaveBeenCalledTimes(2);
         });
 
@@ -100,6 +103,31 @@ describe('Dialogs', () => {
             const result = isFragmentCommandEnabled([]);
 
             expect(result).toBe(false);
+          });
+    });
+
+    describe('isControllerExtensionEnabled', () => {
+        const syncViewsIds = ['syncViewId1', 'syncViewId2'];
+        const elementOverlayMock = { getElement: jest.fn() } as unknown as ElementOverlay;
+
+        it('should return true when overlays length is 1 and clickedControlId is not in syncViewsIds', () => {
+            FlUtils.getViewForControl = jest.fn().mockReturnValue({ getId: jest.fn().mockReturnValue('asyncViewId2') });
+            const overlays: ElementOverlay[] = [elementOverlayMock];
+            expect(isControllerExtensionEnabled(overlays, syncViewsIds)).toBe(true);
+        });
+
+        it('should return false when overlays length is 1 and clickedControlId is in syncViewsIds', () => {
+            FlUtils.getViewForControl = jest.fn().mockReturnValue({ getId: jest.fn().mockReturnValue('syncViewId1') });
+            const overlays: ElementOverlay[] = [elementOverlayMock];
+
+            expect(isControllerExtensionEnabled(overlays, syncViewsIds)).toBe(false);
+        });
+
+        it('should return false when overlays length is more than 1', () => {
+            FlUtils.getViewForControl = jest.fn().mockReturnValue({ getId: jest.fn().mockReturnValue('syncViewId3') });
+            const overlays: ElementOverlay[] = [elementOverlayMock, elementOverlayMock];
+            const syncViewsIds = ['syncViewId1', 'syncViewId2'];
+            expect(isControllerExtensionEnabled(overlays, syncViewsIds)).toBe(false);
         });
     });
 });