+++ packages/preview-middleware-client/types/sap.ui.dt.d.ts
@@ -83,3 +83,14 @@ declare module 'sap/ui/dt/plugin/ContextMenu' {
 
     export default ContextMenu;
 }
+
+declare module 'sap/ui/dt/Element' {
+    import type Element from 'sap/ui/core/Element';
+
+    interface ElementExtended extends Element {
+        oAsyncState: any;
+    }
+
+    export default ElementExtended;
+}
+