+++ packages/create/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/create
 
+## 0.5.76
+
+### Patch Changes
+
+-   @sap-ux/adp-tooling@0.9.28
+-   @sap-ux/app-config-writer@0.3.58
+
+## 0.5.75
+
+### Patch Changes
+
+-   @sap-ux/adp-tooling@0.9.27
+-   @sap-ux/app-config-writer@0.3.57
+
 ## 0.5.74
 
 ### Patch Changes